requirements.txt
requests==2.32.3
bs4
ollama==0.2.1
regex==2023.10.3
joblib==1.3.2
streamlit==1.33.0
pandas==2.1.2
urllib3==2.2.2
python-dotenv==1.0.0
stripe==10.5.0
uvicorn==0.30.1
pymongo
twilio==9.2.3
pyotp==2.9.0
newsapi-python==0.2.7
fastapi==0.111.0
fastapi-cli==0.0.4
fireworks-ai==0.15.1
